+++ src/Execution/Arguments/NestedOneToMany.php
@@ -2,6 +2,8 @@
 
 namespace Nuwave\Lighthouse\Execution\Arguments;
 
+use Closure;
+use Illuminate\Database\Eloquent\Relations\HasOneOrMany;
 use Illuminate\Database\Eloquent\Relations\Relation;
 use Nuwave\Lighthouse\Support\Contracts\ArgResolver;
 
@@ -27,6 +29,7 @@ class NestedOneToMany implements ArgResolver
         $relation = $parent->{$this->relationName}();
 
         static::createUpdateUpsert($args, $relation);
+        static::connectDisconnect($args, $relation);
 
         if ($args->has('delete')) {
             $relation->getRelated()::destroy(
@@ -66,5 +69,50 @@ class NestedOneToMany implements ArgResolver
                 $upsertModel($relation->make(), $childArgs);
             }
         }
+    }
+
+    public static function connectDisconnect(ArgumentSet $args, HasOneOrMany $relation): void
+    {
+        if ($args->has('connect')) {
+            // @phpstan-ignore-next-line Relation&Builder mixin not recognized
+            $children = $relation
+                ->make()
+                ->whereIn(
+                    self::getLocalKeyName($relation),
+                    $args->arguments['connect']->value
+                )
+                ->get();
+
+            // @phpstan-ignore-next-line Relation&Builder mixin not recognized
+            $relation->saveMany($children);
+        }
+
+        if ($args->has('disconnect')) {
+            // @phpstan-ignore-next-line Relation&Builder mixin not recognized
+            $relation
+                ->make()
+                ->whereIn(
+                    self::getLocalKeyName($relation),
+                    $args->arguments['disconnect']->value
+                )
+                ->update([$relation->getForeignKeyName() => null]);
+        }
+    }
+
+    /**
+     * TODO remove this horrible hack when we no longer support Laravel 5.6.
+     */
+    private static function getLocalKeyName(HasOneOrMany $relation): string
+    {
+        $getLocalKeyName = Closure::bind(
+            function () {
+                // @phpstan-ignore-next-line $this variable not recognized despite it's exists in the bind class
+                return $this->localKey;
+            },
+            $relation,
+            get_class($relation)
+        );
+
+        return $getLocalKeyName();
     }
 }