+++ CHANGELOG.md
@@ -12,6 +12,7 @@ You can find and compare releases at the [GitHub release page](https://github.co
 ### Added
 
 - Add `GraphQLContext` to `StartExecution` event
+- Add `connect` and `disconnect` operations in nested mutations for HasMany and MorphMany relations https://github.com/nuwave/lighthouse/pull/1730
 
 ### Changed
 
@@ -21,6 +22,8 @@ You can find and compare releases at the [GitHub release page](https://github.co
 ### Fixed
 
 - Apply error handling and debug settings consistently https://github.com/nuwave/lighthouse/pull/1749
+- Fix typo `comparision` to `comparison` in generated input types for `@whereHas`
+- Fix redis `mget` being called with an empty list of subscriber ids https://github.com/nuwave/lighthouse/pull/1759
 
 ## 5.3.0